# web/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BeautyMS API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/master/register": {
            "post": {
                "tags": [
                    "Master"
                ],
                "summary": "Регистрация мастера",
                "operationId": "b06a7c51e80c1818756f2932db3285be",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "middlename": {
                                        "type": "string"
                                    },
                                    "firstname": {
                                        "type": "string"
                                    },
                                    "birthday": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "gender": {
                                        "type": "integer"
                                    },
                                    "client_gender": {
                                        "type": "integer"
                                    },
                                    "work_city": {
                                        "type": "string"
                                    },
                                    "work_street": {
                                        "type": "string"
                                    },
                                    "work_house": {
                                        "type": "string"
                                    },
                                    "live_city": {
                                        "type": "string"
                                    },
                                    "live_street": {
                                        "type": "string"
                                    },
                                    "live_house": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "URL для оплаты",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "url": {
                                            "type": "string",
                                            "format": "uri"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка запроса"
                    },
                    "401": {
                        "description": "Ошибка аутентификации"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/master/profile": {
            "get": {
                "tags": [
                    "Master"
                ],
                "summary": "Получение профиля мастера",
                "operationId": "218a61f3f5281aa2955c4848ec778c7b",
                "responses": {
                    "200": {
                        "description": "Профиль мастера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "middlename": {
                                            "type": "string"
                                        },
                                        "firstname": {
                                            "type": "string"
                                        },
                                        "birthday": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "date": {
                                            "type": "integer"
                                        },
                                        "balance": {
                                            "type": "number"
                                        },
                                        "gender": {
                                            "type": "integer"
                                        },
                                        "client_gender": {
                                            "type": "integer"
                                        },
                                        "status": {
                                            "type": "integer"
                                        },
                                        "search_radius": {
                                            "type": "integer"
                                        },
                                        "work_city": {
                                            "type": "string"
                                        },
                                        "work_street": {
                                            "type": "string"
                                        },
                                        "work_house": {
                                            "type": "string"
                                        },
                                        "work_lat": {
                                            "type": "number"
                                        },
                                        "work_lon": {
                                            "type": "number"
                                        },
                                        "live_city": {
                                            "type": "string"
                                        },
                                        "live_street": {
                                            "type": "string"
                                        },
                                        "live_house": {
                                            "type": "string"
                                        },
                                        "live_lat": {
                                            "type": "number"
                                        },
                                        "live_lon": {
                                            "type": "number"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка аутентификации"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/master/update": {
            "post": {
                "tags": [
                    "Master"
                ],
                "summary": "Обновление профиля мастера",
                "operationId": "383c63d9e110a1a79126742de63947f4",
                "responses": {
                    "200": {
                        "description": "Обновленный профиль пользователя",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка аутентификации"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/master/balance": {
            "post": {
                "tags": [
                    "Master"
                ],
                "summary": "Пополнение баланса",
                "operationId": "e8c2762d18f79daca89d585fbd61e78c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "price": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "URL для оплаты",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "url": {
                                            "type": "string",
                                            "format": "uri"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка запроса"
                    },
                    "401": {
                        "description": "Ошибка аутентификации"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}